Movido la creacion de ruta donde deveria estar, deleteado carpeta innecesaria.
